검색 / css 수정 / 지도 분리작업 진행중
