+++ front/package.json
@@ -9,7 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
     "@googlemaps/react-wrapper": "^1.1.42",
+    "@mui/icons-material": "^6.3.0",
     "@types/google.maps": "^3.58.1",
     "@vis.gl/react-google-maps": "^1.4.2",
     "next": "15.1.1",
