+++ front/src/app/scheduel/page.jsx
@@ -1,0 +1,4 @@
+
+export default function Map() {
+    return <h1>스케줄</h1>
+}
