+++ front/src/context/MapContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react"
+
+export const MapContext = createContext(null);
