+++ front/src/styles/map.module.css
@@ -3,15 +3,50 @@
     margin-top: 30px;
 }
 .container > div:first-child {
+    overflow: hidden;
+    display: flex;
+    border-right: 3px solid rgba(0, 0, 0, 0.25) ;
     padding: 10px;
-    width: 40%;
+    width: 50%;
+    justify-content: center;
+    flex-direction: column;
 }
 
 .container > div:nth-child(2) {
-    margin:0 10px ;
+    margin:0 10px;
     width: 40%;
     padding: 10px;
 }
+
+.container > div:nth-child(2) > button {
+    margin-left: 218px; margin-top: 2px;
+    border-radius: 5px;
+    padding: 5px 10px;
+    width: 100px;
+    font-size: 16px;
+    font-weight: 700;
+    border: 1px solid rgba(0, 0, 0, 0.25);
+    background-color: white;
+}
+
+.container > div:nth-child(2) > button:hover {
+    cursor: pointer;
+    background-color: #81BFDA;
+    border: 2px solid rgba(64, 99, 114);
+    transition: 0.3s;
+}
+.container > div:nth-child(2) > button.select {
+    transition: 0.2s;
+    background-color: #81BFDA;
+    border: 2px solid rgba(64, 99, 114);
+    font-weight: 700;
+}
+
+.container > div:nth-child(2) > button.select:hover {
+    border: 1px solid rgba(0, 0, 0, 0.25);
+    background-color: white;
+}
+
 .subtitle {
     position: sticky;
     margin: 10px 0 ; margin-left: 20px;
@@ -20,7 +55,7 @@
 }
 
 .address {
-    font-size: 12px;
+    font-size: 14px;
 }
 
 .text {
@@ -28,13 +63,65 @@
     font-weight: 700;
 }
 
-.container > div > div > button {
-    margin: 10px 0;
+.cate {
+    width: 100%;
+}
+
+.cate > button {
+    font-size: 1.01rem;
+    border-radius: 8px;
+    border: 1px solid rgba(0, 0, 0, 0.25);
+    background-color: white;
+    margin: 20px 0 10px 0;
+    margin-left: 5px;
+    width : 100px;
+    padding: 8px 0;
+    transition: 0.3s;
+}
+.cate > button:first-child {
+    margin-left: 15px;
+}
+.cate > button:hover {
+    cursor: pointer;
+    background-color: #81BFDA;
+    border: 2px solid rgba(58, 64, 97);
+    transition: 0.3s;
+}
+.cate > button.select {
+    transition: 0.2s;
+    background-color: #81BFDA;
+    border: 2px solid rgba(58, 64, 97);
+    font-weight: 700;
+}
+.searchBar {
     margin-left: 20px;
-    padding: 5px 10px;
+}
+.searchBar > input {
+    outline: none;
+    border-radius: 5px;
+    border-color: rgba(0, 0, 0, 0.25);
+    width: 93%;
+    height: 2rem;
+    font-size: 1rem;
+    padding : 5px 15px;
+}
+.searchBar > button {
+    border: none;
+    background: white;
+    padding: 0;
+}
+
+.searchBar > button > svg {
+    position: absolute; top: 168px; left: 23vw;
+    width: 36px; height: 36px;
+}
+
+.searchBar > button:hover {
+    cursor: pointer;
 }
 
 .search {
+    box-sizing: border-box;
     padding: 10px 20px;
     width: 100%;
     height: 100vh;
@@ -51,8 +138,8 @@
     display: flex;
     padding: 10px;
     border: 1px solid rgba(0, 0, 0, 0.25); border-radius: 5px;
-    box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.25);
-    margin: 10px 0;
+    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
+    margin: 30px 0;
     align-items: center;
 }
 
@@ -78,13 +165,34 @@
 }
 
 .search button {
+    font-size: 0.9rem;
+    background-color: white;
+    border: 1px solid rgba(0, 0, 0, 0.25);
     text-align: center;
     position: relative;
     border-radius: 5px;
-    padding: 5px 12px;
-    left: 2%; top: 20px;
+    width: 100px;
+    padding: 5px 25px;
+    left: 0%; top: 20px;
+    transition: 0.2s;
+}
+.search button.select{
+    background-color: #213555;
+    color: white;
 }
 
+.search button.select:hover{
+    background-color: white;
+    border: 1px solid rgba(0, 0, 0, 0.25);
+    color: black;
+}
+
+.search button:hover {
+    cursor: pointer;
+    background-color: #213555;
+    color: white;
+    transition: 0.2s;
+}
 
 .selectBox {
     overflow: scroll;
@@ -115,7 +223,7 @@
     border-radius: 5px;
 }
 
-.selectBox > div > div{
+.selectBox > div > div {
     margin: 0 10px;
 }
 .selectBox > div > div:nth-child(2) {
@@ -126,8 +234,16 @@
 }
 
 .selectBox button {
-    text-align: center;
     position: relative;
-    border-radius: 5px;
-    padding: 5px 12px;
+    background-color: white;
+    border: none;
+}
+.selectBox button > svg {
+    width: 30px; height: 30px;
+}
+
+.selectBox button > svg:hover {
+    cursor: pointer;
+    color: #d80400;
+    transition: 0.2s;
 }
